requirements.txt
streamlit==1.44.1
torch==2.6.0
torchvision==0.21.0
matplotlib==3.10.1
Pillow==11.2.1
numpy==2.2.5
scikit-learn==1.6.1
seaborn==0.13.2
opencv-python==4.11.0.86
